+++ tests/telemetry/test_events.py
@@ -27,18 +27,9 @@ def validate_yang(duthost, op_file="", yang_file=""):
     assert ret["rc"] == 0, "Yang validation failed for {}".format(yang_file)
 
 
-def do_init(duthost):
-    for i in [BASE_DIR, DATA_DIR]:
-        try:
-            os.mkdir(i)
-        except OSError as e:
-            logger.info("Dir/file already exists: {}, skipping mkdir".format(e))
-
-    duthost.copy(src="telemetry/validate_yang_events.py", dest="/tmp")
-
-
 @pytest.mark.disable_loganalyzer
-def test_events(duthosts, enum_rand_one_per_hwsku_hostname, ptfhost, setup_streaming_telemetry, localhost, gnxi_path):
+def test_events(duthosts, enum_rand_one_per_hwsku_hostname, ptfhost, setup_streaming_telemetry, localhost, gnxi_path,
+                test_eventd_healthy):
     """ Run series of events inside duthost and validate that output is correct
     and conforms to YANG schema
     """
@@ -46,11 +37,10 @@ def test_events(duthosts, enum_rand_one_per_hwsku_hostname, ptfhost, setup_strea
     logger.info("Start events testing")
 
     skip_201911_and_older(duthost)
-    do_init(duthost)
 
-    # Load all events test code and run
+    # Load rest of events
     for file in os.listdir(EVENTS_TESTS_PATH):
-        if file.endswith("_events.py"):
+        if file.endswith("_events.py") and not file.endswith("eventd_events.py"):
             module = __import__(file[:len(file)-3])
             module.test_event(duthost, gnxi_path, ptfhost, DATA_DIR, validate_yang)
             logger.info("Completed test file: {}".format(os.path.join(EVENTS_TESTS_PATH, file)))
